+++ packages/core-web-generator/package.json
@@ -3,7 +3,11 @@
   "version": "0.4.0",
   "private": true,
   "scripts": {
-    "build": "node ./generate.js"
+    "build": "node ./generate.js",
+    "test": "node ./__tests__/core-web-generator.test.js"
+  },
+  "directories": {
+    "test": "__tests__"
   },
   "dependencies": {
     "@mdn/browser-compat-data": "^3.0.1",
